// src/entities/product/model/types.ts
export interface Product {
    id: number;
    title: string;
    description: string;
    category: string;
    price: number;
    discountPercentage: number;
    rating: number;
    stock: number;
    brand: string;
    sku: string;
    thumbnail: string;
    images: string[];
}

export interface ProductsResponse {
    products: Product[];
    total: number;
    skip: number;
    limit: number;
}

export interface ProductsParams {
    limit?: number;
    skip?: number;
    sortBy?: string;
    order?: "asc" | "desc";
}

export interface AddProductDto {
    title: string;
    price: number;
    brand: string;
    sku: string;
}